recolector/requirements.txt
# recolector/requirements.txt
selenium==4.11.2
python-dotenv==1.0.1